study_buddy/study_buddy/requirements.txt
crewai[tools]
PyPDF2
agentops
python-multipart
asyncpg
passlib
python-jose[cryptography]
celery[redis]
aioredis
prometheus-fastapi-instrumentator